% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Example.R
\encoding{UTF-8}
\name{Example}
\alias{Example}
\title{Arquivos para Exemplificar o uso das Funções}
\usage{
Example()
}
\value{
Essa função retorna uma \code{lista} contendo 3 \code{data.frame}:
\itemize{
\item '\code{Dados}': dados fictícios.
\item \code{Label}: dicionário dos dados fictícios.
\item \code{Enunciado}: enunciado dos dados fictícios.
}
}
\description{
Essa função gera três data.frames's que podem ser usados para testar e usar as funções do \code{PacoteInnovare}.
}
\examples{
#Executando
Dados<-Example()[[1]]
Label<-Example()[[2]]
Enunciado<-Example()[[3]]
#Verificando
head(Dados)
head(Label)
head(Enunciado)


}
\author{
Felipe Ferreira \email{felipe179971@hotmail.com} e Rafael Peixoto \email{peixoto.rafael@outlook.com}.
}
