% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUN_isoladas.R
\encoding{UTF-8}
\name{FUN_isoladas}
\alias{FUN_isoladas}
\title{Função de Frequência}
\usage{
FUN_isoladas(TABELA, DICIONARIO, variaveis, adc_labels = T)
}
\arguments{
\item{TABELA}{Tabela já filtrada pelos splits}

\item{DICIONARIO}{Arquivo com o significado dos labels.}

\item{variaveis}{Quais variáveis serão calculadas.}

\item{adc_labels}{se TRUE (o padrão) retorna uma coluna conteno o label do split}
}
\value{
Se \code{adc_labels=TRUE}, essa função retorna um \code{tibble} contendo 6 colunas:
\itemize{
\item '\code{variável}': valores possíveis da variável.
\item \code{n}: número de observações da \code{variaveis}.
\item \code{n_peso}: número de observações da \code{variaveis} (com peso aplicado).
\item \code{pct}: porcentagem da \code{variaveis}.
\item \code{pct_peso}: porcentagem da \code{variaveis} (com peso aplicado).
\item \code{variável_label}: Retorna o label do split (só se \code{adc_labels=TRUE}).
}
}
\description{
Essa função calcula a frequência e a porcentagem das variáveis desejadas.
}
\examples{
library(tidyverse)
#Criando o data.frame com os dados#########################
#Número de linhas
n=100
#Variável de sexo
set.seed(1)
v1=sample(1:2,n,replace = TRUE)
#variável de renda
set.seed(2)
v2=sample(1:4,n,replace = TRUE)
#Variável da UF
set.seed(3)
v3=sample(1:10,n,replace = TRUE)
#Peso (sexo)
v4<-v1
v4[which(v4==1)]<-(n/2)/length(v1[which(v1==1)])
v4[which(v4==2)]<-(n/2)/length(v1[which(v1==2)])

dados<-data.frame(v1,v2,v3,v4)
#Criando o Dicionário dos dados#########################
Label=data.frame(
  id=c(rep(1,length(v1\%>\%unique())),
       rep(2,length(v2\%>\%unique())),
       rep(3,length(v3\%>\%unique()))),
  variavel=c(rep("v1",length(v1\%>\%unique())),
             rep("v2",length(v2\%>\%unique())),
             rep("v3",length(v3\%>\%unique()))),
  codigo=c(v1\%>\%unique(),v2\%>\%unique(),v3\%>\%unique()),
  label=c("Masculino","Feminino",
          "Baixa Renda","Classe Média","Classe Alta","Milionário",
          paste("Cidade",seq(1:10)))
)
#Tratando para entrar na função#######################
dados_peso <- dados \%>\%
  dplyr::mutate("peso" = v4)

splits <- dados_peso \%>\% dplyr::mutate(v1=as.character(v1))\%>\%
  dplyr::distinct(v1) \%>\%
  dplyr::left_join(
    y = Label \%>\%
      dplyr::filter(variavel == "v1") \%>\%
      dplyr::select("v1" = codigo, "splits" = label) \%>\%
      dplyr::mutate(across(v1, as.character))
  )
Dados_splits <- dados_peso \%>\% dplyr::mutate(v1=as.character(v1))\%>\%
  dplyr::left_join(splits) \%>\%
  dplyr::group_nest(splits) \%>\%
  dplyr::bind_rows(
    dados_peso \%>\% dplyr::mutate(v1=as.character(v1))\%>\%
      dplyr::mutate(splits = "Geral") \%>\%
      dplyr::group_nest(splits)
  ) \%>\%
  dplyr::mutate(data = map(data, ~ dplyr::select(., peso, matches("v\\\\d+"))))
Label<-Label\%>\%dplyr::rename(opcao_variavel="variavel",opcao_cod="codigo",opcao_label="label")

#Executando#########################################
#Apenas uma variável sem label do split (adc_labels = FALSE)
FUN_isoladas(TABELA = dados_peso, DICIONARIO = Label, variaveis = "v1",
            adc_labels = FALSE)
#Mais de uma variável e com label do split (adc_labels = TRUE)
FUN_isoladas(TABELA = dados_peso, DICIONARIO = Label, variaveis = c("v1","v2","v3"),
             adc_labels = TRUE)
#Pelos splits, com label
Dados_splits\%>\%
  dplyr::mutate(data = map(data, .f = ~ FUN_isoladas(., DICIONARIO = Label, variaveis = "v1",
   adc_labels = TRUE)))

rm(a,dados,dados_peso,Dados_splits,Label,splits,n,v1,v2,v3,v4)\%>\%suppressWarnings()

}
\author{
Felipe Ferreira \email{felipe179971@hotmail.com} e Rafael Peixoto \email{peixoto.rafael@outlook.com}.
}
