% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUN_media.R
\encoding{UTF-8}
\name{FUN_media}
\alias{FUN_media}
\title{Função de Frequência}
\usage{
FUN_media(TABELA, variaveis, vars_na = c(-88, -99), aparada = FALSE, cut = 2.5)
}
\arguments{
\item{TABELA}{Tabela já filtrada pelos splits.}

\item{variaveis}{Variáveis cuja média deverá ser calculada.}

\item{vars_na}{Código a serem transformados em NA .}

\item{aparada}{Se quer (TRUE) ou não (FALSE) a média aparada.}

\item{cut}{Para a média aparada: Porcentagem de corte (Ex.: cut=2.5 para remover os 2.5\% menores e 97.5\% maiores)}
}
\value{
Se \code{aparada=TRUE}, essa função retorna um \code{tibble} contendo 6 colunas:
\itemize{
\item '\code{variavel}': ...
\item \code{media}: ...
\item \code{media_peso}: ...
\item \code{n}: ...
\item \code{n_peso}: ...
\item \code{desvp}: ...
\item \code{erro}: ...
\item \code{erro_peso}: ...
}
}
\description{
Função apra calcular a média e/ou a média aparada.
}
\details{
\itemize{
\item \code{TABELA}: o parâmetro só aceita tabelas no formato \code{Nest Tibble} (splits  x data). Além disso, precisa ter uma coluna chamada “peso” contendo o peso a ser aplicado ao respectivo Nest. Caso nenhum peso precise ser aplicado, atribuir peso=1.

}
}
\examples{

library(tidyverse)
#Arquivos
dados<-PacoteInnovare::Example()[[1]]
Label<-PacoteInnovare::Example()[[2]]
Enunciado<-PacoteInnovare::Example()[[3]]
##Fazendo o tratamento para entrar na função
#Como tenho dois arquivos, um para o enunciado e outro para o dicionário,
#usarei a função func_labels
Dicionario=PacoteInnovare::func_labels(tabela_labels=Label,
                                       tabela_enunciado=Enunciado,
                                       col_codigo="codigo",
                                       col_label="label",
                                       col_var="variavel",
                                       col_enunciado="enunciado")
#Identificando nosso splits
split_sexo<-Dicionario\%>\%
  dplyr::filter(opcao_variavel=="v1")\%>\%
  dplyr::select("v1"=opcao_cod,"splits"=opcao_label)
split_time<-Dicionario\%>\%
  dplyr::filter(opcao_variavel=="v17")\%>\%
  dplyr::select("v17"=opcao_cod,"splits"=opcao_label)
split_tipo<-Dicionario\%>\%
  dplyr::filter(opcao_variavel=="v19")\%>\%
  dplyr::select("v19"=opcao_cod,"splits"=opcao_label)
#Criando a tabela no formato Nest Tibble dividido pelos splits
dados_split<-dplyr::bind_rows(
  #Split por sexo, cujo peso está na v4
  dados \%>\%
    # Atribuir o peso para a coluna "peso"
    dplyr::mutate("peso" = v4) \%>\%
    # Juntar a tabela com os labels dos splits
    dplyr::left_join(split_sexo) \%>\%
    # dividir a tabela pelos splits
    dplyr::group_nest(splits),
  #Adicionando o Split por time, cujo peso está na v18
  dados \%>\%
    # Atribuir o peso para a coluna "peso"
    dplyr::mutate("peso" = v18) \%>\%
    # Juntar a tabela com os labels dos splits
    dplyr::left_join(split_time) \%>\%
    # dividir a tabela pelos splits
    dplyr::group_nest(splits),
  #Adicionando o Split por tipo (não tem peso)
  dados \%>\%
    # Atribuir o peso para a coluna "peso"
    dplyr::mutate("peso" = 1) \%>\%
    # Juntar a tabela com os labels dos splits
    dplyr::left_join(split_tipo) \%>\%
    # dividir a tabela pelos splits
    dplyr::group_nest(splits),
  #Adicionando o Geral (não tem peso)
  dados\%>\%
    # Atribuir o peso para a coluna "peso"
    dplyr::mutate("peso" = 1,splits=as.character("Geral")) \%>\%
    # dividir a tabela pelos splits
    dplyr::group_nest(splits)
)
##Calculando a média de v15,v16 e v21 para o split Feminino
#(utilziando o padrão de vars_na, aparada e cut)
PacoteInnovare::FUN_media(TABELA=dados_split[[2]][[which(dados_split$splits=="Geral")]],
                          variaveis=c("v15","v16","v21"))
#Transformando apenas o código -88 em NA
PacoteInnovare::FUN_media(TABELA=dados_split[[2]][[which(dados_split$splits=="Geral")]],
                          variaveis=c("v15","v16","v21"),
                          vars_na=c(-88)
                          )
#Média aparada (2.5\%) e nenhum código sendo transformando em NA
PacoteInnovare::FUN_media(TABELA=dados_split[[2]][[which(dados_split$splits=="Geral")]],
                          variaveis=c("v15","v16","v21"),
                          vars_na=NA,
                          aparada = TRUE
)
#Média aparada (3\%) com -88,-99 sendo transformando em NA
PacoteInnovare::FUN_media(TABELA=dados_split[[2]][[which(dados_split$splits=="Geral")]],
                          variaveis=c("v15","v16","v21"),
                          aparada = TRUE,
                          cut=3
)

#Observando alguns erros e warnings da função
dados_split[[2]][[which(dados_split$splits=="Geral")]]\%>\%dplyr::mutate(
  v15=NA,
  v16=ifelse(v16==6,"character",v16),
  v21=as.character(v21)
)
PacoteInnovare::FUN_media(
  TABELA=dados_split[[2]][[which(dados_split$splits=="Geral")]]\%>\%dplyr::mutate(
    v15=NA,
    v16=ifelse(v16==6,"character",v16),
    v21=as.character(v21)
    ),
  variaveis=c("v15","v16","v21"),
)
#Erro CUT
PacoteInnovare::FUN_media(TABELA=dados_split[[2]][[which(dados_split$splits=="Geral")]],
                          variaveis=c("v15","v16","v21"),
                          cut=50)




}
\author{
Felipe Ferreira \email{felipe179971@hotmail.com} e Rafael Peixoto \email{peixoto.rafael@outlook.com}.
}
