% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUN_MRG.R
\encoding{UTF-8}
\name{FUN_MRG}
\alias{FUN_MRG}
\title{Função de Frequência}
\usage{
FUN_MRG(TABELA, DICIONARIO, lista_variaveis, adc_labels = T)
}
\arguments{
\item{TABELA}{Tabela já filtrada pelos splits.}

\item{DICIONARIO}{Arquivo com o significado dos labels.}

\item{lista_variaveis}{Quais MRG's serão calculados.}

\item{adc_labels}{se TRUE (o padrão) retorna uma coluna conteno o label do split.}
}
\value{
Se \code{adc_labels=TRUE}, essa função retorna um \code{tibble} contendo 6 colunas:
\itemize{
\item '\code{lista_variaveis}': valores possíveis da variável;
\item \code{n}: número de observações da \code{lista_variaveis};
\item \code{n_peso}: número de observações da \code{variaveis} (com peso aplicado);
\item \code{pct}: porcentagem da \code{lista_variaveis};
\item \code{pct_peso}: porcentagem da \code{lista_variaveis} (com peso aplicado);
\item \code{variável_label}: retorna o label do split (só se \code{adc_labels=TRUE}).
}
}
\description{
Essa função calcula a frequência e a porcentagem das variáveis MRG. Para executar MRG do tipo citou/não citou, consultar \code{FUN_Citou}.
}
\details{
\itemize{
\item \code{TABELA}: o parâmetro só aceita tabelas no formato \code{Nest Tibble} (splits  x data). Além disso, precisa ter uma coluna chamada “peso” contendo o peso a ser aplicado ao respectivo Nest. Caso nenhum peso precise ser aplicado, atribuir peso=1.
\item \code{DICIONARIO}: o arquivo do dicionário precisa, OBRIGATORIAMENTE, ter no mínimo as seguintes colunas:
\itemize{
\item \code{opcao_cod}: coluna contendo o código.
\item \code{opcao_label}: coluna contendo o significado do código;
\item \code{opcao_variavel}: coluna contendo a variável;
\item \code{pergunta_enunciado}: coluna contendo o enunciado da vaiável.

\itemize{
\item{\code{Caso você tenha dois arquivos separador, sendo um para os labels e outro para o enunciado, ver func_labels().}}
}
\item\code{lista_variaveis}: lista contendo os MRG’s. Todos os MRG’s especificados nesse parâmetro precisam, obrigatoriamente, possuir seu respectivo label no arquivo especificado em \code{DICIONARIO} no formato "primeira variável""mrg". Ex.: mrg=list("G013"=c("v13","v14")) sendo que existe um v13mrg no arquivo especificado em \code{DICIONARIO}).
}

}
}
\examples{

library(tidyverse)
#Arquivos
dados<-PacoteInnovare::Example()[[1]]
Label<-PacoteInnovare::Example()[[2]]
Enunciado<-PacoteInnovare::Example()[[3]]
##Fazendo o tratamento para entrar na função
#Como tenho dois arquivos, um para o enunciado e outro para o dicionário,
#usarei a função func_labels
Dicionario=PacoteInnovare::func_labels(tabela_labels=Label,
                                       tabela_enunciado=Enunciado,
                                       col_codigo="codigo",
                                       col_label="label",
                                       col_var="variavel",
                                       col_enunciado="enunciado")
#Identificando nosso splits
split_sexo<-Dicionario\%>\%
  dplyr::filter(opcao_variavel=="v1")\%>\%
  dplyr::select("v1"=opcao_cod,"splits"=opcao_label)
split_time<-Dicionario\%>\%
  dplyr::filter(opcao_variavel=="v17")\%>\%
  dplyr::select("v17"=opcao_cod,"splits"=opcao_label)
split_tipo<-Dicionario\%>\%
  dplyr::filter(opcao_variavel=="v19")\%>\%
  dplyr::select("v19"=opcao_cod,"splits"=opcao_label)
#Criando a tabela no formato Nest Tibble dividido pelos splits
dados_split<-dplyr::bind_rows(
  #Split por sexo, cujo peso está na v4
  dados \%>\%
    # Atribuir o peso para a coluna "peso"
    dplyr::mutate("peso" = v4) \%>\%
    # Juntar a tabela com os labels dos splits
    dplyr::left_join(split_sexo) \%>\%
    # dividir a tabela pelos splits
    dplyr::group_nest(splits),
  #Adicionando o Split por time, cujo peso está na v18
  dados \%>\%
    # Atribuir o peso para a coluna "peso"
    dplyr::mutate("peso" = v18) \%>\%
    # Juntar a tabela com os labels dos splits
    dplyr::left_join(split_time) \%>\%
    # dividir a tabela pelos splits
    dplyr::group_nest(splits),
  #Adicionando o Split por tipo (não tem peso)
  dados \%>\%
    # Atribuir o peso para a coluna "peso"
    dplyr::mutate("peso" = 1) \%>\%
    # Juntar a tabela com os labels dos splits
    dplyr::left_join(split_tipo) \%>\%
    # dividir a tabela pelos splits
    dplyr::group_nest(splits),
  #Adicionando o Geral (não tem peso)
  dados\%>\%
    # Atribuir o peso para a coluna "peso"
    dplyr::mutate("peso" = 1,splits=as.character("Geral")) \%>\%
    # dividir a tabela pelos splits
    dplyr::group_nest(splits)
)
#Calculando a frequência do MRG das variáveis v13 e v14 no split "Geral" com label
PacoteInnovare::FUN_MRG(TABELA=dados_split[[2]][[which(dados_split$splits=="Geral")]],
                        DICIONARIO=Dicionario,
                        lista_variaveis=list("MRG13"=c("v13","v14")),
                        adc_labels = TRUE)
#Calculando a frequência do MRG das variáveis v13 e v14 no split "Feminino" sem label
PacoteInnovare::FUN_MRG(TABELA=dados_split[[2]][[which(dados_split$splits=="Feminino")]],
                        DICIONARIO=Dicionario,
                        lista_variaveis=list("MRG13"=c("v13","v14")),
                        adc_labels = FALSE)


}
\author{
Felipe Ferreira \email{felipe179971@hotmail.com} e Rafael Peixoto \email{peixoto.rafael@outlook.com}.
}
