% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUN_isoladas.R
\encoding{UTF-8}
\name{FUN_isoladas}
\alias{FUN_isoladas}
\title{Função de Frequência}
\usage{
FUN_isoladas(TABELA, DICIONARIO, variaveis, adc_labels = TRUE)
}
\arguments{
\item{TABELA}{Tabela já filtrada pelos splits}

\item{DICIONARIO}{Arquivo com o significado dos labels.}

\item{variaveis}{Quais variáveis serão calculadas.}

\item{adc_labels}{se TRUE (o padrão) retorna uma coluna conteno o label do split}
}
\value{
Se \code{adc_labels=TRUE}, essa função retorna um \code{tibble} contendo 6 colunas:
\itemize{
\item '\code{variável}': valores possíveis da variável.
\item \code{n}: número de observações da \code{variaveis}.
\item \code{n_peso}: número de observações da \code{variaveis} (com peso aplicado).
\item \code{pct}: porcentagem da \code{variaveis}.
\item \code{pct_peso}: porcentagem da \code{variaveis} (com peso aplicado).
\item \code{variável_label}: Retorna o label do split (só se \code{adc_labels=TRUE}).
}
}
\description{
Essa função calcula a frequência e a porcentagem das variáveis desejadas.
}
\examples{

library(tidyverse)
#Arquivos
dados<-PacoteInnovare::Example()[[1]]
Label<-PacoteInnovare::Example()[[2]]
Enunciado<-PacoteInnovare::Example()[[3]]
##Fazendo o tratamento para entrar na função
#Como tenho dois arquivos, um para o enunciado e outro para o dicionário,
#usarei a função func_labels
Dicionario=PacoteInnovare::func_labels(tabela_labels=Label,
                                       tabela_enunciado=Enunciado,
                                       col_codigo="codigo",
                                       col_label="label",
                                       col_var="variavel",
                                       col_enunciado="enunciado")
#Identificando nosso splits
split_sexo<-Dicionario\%>\%
  dplyr::filter(opcao_variavel=="v1")\%>\%
  dplyr::select("v1"=opcao_cod,"splits"=opcao_label)
split_time<-Dicionario\%>\%
  dplyr::filter(opcao_variavel=="v17")\%>\%
  dplyr::select("v17"=opcao_cod,"splits"=opcao_label)
split_tipo<-Dicionario\%>\%
  dplyr::filter(opcao_variavel=="v19")\%>\%
  dplyr::select("v19"=opcao_cod,"splits"=opcao_label)
#Criando a tabela no formato Nest Tibble dividido pelos splits
dados_split<-dplyr::bind_rows(
  #Split por sexo, cujo peso está na v4
  dados \%>\%
    # Atribuir o peso para a coluna "peso"
    dplyr::mutate("peso" = v4) \%>\%
    # Juntar a tabela com os labels dos splits
    dplyr::left_join(split_sexo) \%>\%
    # dividir a tabela pelos splits
    dplyr::group_nest(splits),
  #Adicionando o Split por time, cujo peso está na v18
  dados \%>\%
    # Atribuir o peso para a coluna "peso"
    dplyr::mutate("peso" = v18) \%>\%
    # Juntar a tabela com os labels dos splits
    dplyr::left_join(split_time) \%>\%
    # dividir a tabela pelos splits
    dplyr::group_nest(splits),
  #Adicionando o Split por tipo (não tem peso)
  dados \%>\%
    # Atribuir o peso para a coluna "peso"
    dplyr::mutate("peso" = 1) \%>\%
    # Juntar a tabela com os labels dos splits
    dplyr::left_join(split_tipo) \%>\%
    # dividir a tabela pelos splits
    dplyr::group_nest(splits),
  #Adicionando o Geral (não tem peso)
  dados\%>\%
    # Atribuir o peso para a coluna "peso"
    dplyr::mutate("peso" = 1,splits=as.character("Geral")) \%>\%
    # dividir a tabela pelos splits
    dplyr::group_nest(splits)
)
#Calculando a frequência de "v1" no split "Geral"
PacoteInnovare::FUN_isoladas(TABELA=dados_split[[2]][[which(dados_split$splits=="Geral")]],
                             DICIONARIO=Dicionario,
                             variaveis="v1",
                             adc_labels = TRUE)


}
\author{
Felipe Ferreira \email{felipe179971@hotmail.com} e Rafael Peixoto \email{peixoto.rafael@outlook.com}.
}
