% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_labels.R
\encoding{UTF-8}
\name{func_labels}
\alias{func_labels}
\title{func_labels}
\usage{
func_labels(
  tabela_labels,
  tabela_enunciado,
  col_codigo,
  col_label,
  col_var,
  col_enunciado
)
}
\arguments{
\item{tabela_labels}{...}

\item{tabela_enunciado}{...}

\item{col_codigo}{...}

\item{col_label}{...}

\item{col_var}{...}

\item{col_enunciado}{...}
}
\value{
...
}
\description{
Calcula a diferença entre dois vetores do tipo \code{Sys.time}.
Normalmente utilizada para avaliar o tempo de execução das demais funções do
\code{PaceteInnovare}, é internalizada em diversas funções do pacote mas pode ser utilizada livremente.
}
\author{
Felipe Ferreira \email{felipe179971@hotmail.com} e Rafael Peixoto \email{peixoto.rafael@outlook.com}.
}
